add events module.  move ledger.list -> events.list, ledger.transactions -> events.info
