+++ src/network/modules/events/index.ts
@@ -1,0 +1,14 @@
+export { Events, BoundType, OrderType } from "./events"
+export type {
+  ListFilterArgs,
+  Event,
+  SendEvent,
+  EventsListResponse,
+  CreateAccountEvent,
+  MultisigEvent,
+  MultisigSubmitEvent,
+  MultisigApproveEvent,
+  MultisigExecuteEvent,
+  MultisigRevokeEvent,
+  MultisigWithdrawEvent,
+} from "./events"
