+++ src/network/modules/index.ts
@@ -4,3 +4,4 @@ export * from "./blockchain"
 export * from "./types"
 export * from "./id-store"
 export * from "./account"
+export * from "./events"
