+++ src/network/modules/ledger/__tests__/ledger.test.ts
@@ -1,30 +1,13 @@
-import {
-  getLedgerInfo,
-  getBalance,
-  Ledger,
-  makeListFilters,
-  OrderType,
-  BoundType,
-  setRangeBound,
-  RangeType,
-} from "../ledger"
+import { getLedgerInfo, getBalance, Ledger } from "../ledger"
 import {
   expectedBalancesMap,
   expectedSymbolsMap,
   mockLedgerInfoResponseMessage,
   mockLedgerBalanceResponseMessage,
-  mockLedgeListResponseMessage,
-  expectedListResponse,
-  mockLedgerListMultisigSubmitTxnResponse,
-  expectedMockLedgerListMultisigSubmitTxnResponse,
-  mockLedgerListMultisigTxnsResponse,
-  expectedMockLedgerListMultisigTxnsResponse,
-  expectedMockLedgerListAccountCreateResponse,
-  mockLedgerListAccountCreateResponse,
 } from "./data"
 
 describe("Ledger", () => {
-  describe("ledger.info()", () => {
+  describe("info()", () => {
     it("calling ledger.info() return symbols", async () => {
       const mockCall = jest.fn(async () => {
         return mockLedgerInfoResponseMessage
@@ -41,7 +24,7 @@ describe("Ledger", () => {
     })
   })
 
-  describe("ledger.balance()", () => {
+  describe("balance()", () => {
     it("should call ledger.balance() with an empty list of token identities", async () => {
       const mockCall = jest.fn(async () => {
         return mockLedgerBalanceResponseMessage
@@ -70,160 +53,6 @@ describe("Ledger", () => {
     it("getBalance() should return balances", async () => {
       const res = getBalance(mockLedgerBalanceResponseMessage)
       expect(res).toEqual(expectedBalancesMap)
-    })
-  })
-
-  describe("ledger.list", () => {
-    it("should return count and list of send transactions", async () => {
-      const mockCall = jest.fn(async () => {
-        return mockLedgeListResponseMessage
-      })
-      const ledger = setupLedger(mockCall)
-      const res = await ledger.list()
-      expect(res).toEqual(expectedListResponse)
-    })
-    it("should return count and list of multisigSubmit transaction", async () => {
-      const mockCall = jest.fn(async () => {
-        return mockLedgerListMultisigSubmitTxnResponse
-      })
-      const ledger = setupLedger(mockCall)
-      const res = await ledger.list()
-      expect(res).toEqual(expectedMockLedgerListMultisigSubmitTxnResponse)
-    })
-    it("should return multisig approve, revoke, execute, withdraw transactions", async function () {
-      const mockCall = jest.fn(async () => {
-        return mockLedgerListMultisigTxnsResponse
-      })
-      const ledger = setupLedger(mockCall)
-      const res = await ledger.list()
-      expect(res).toEqual(expectedMockLedgerListMultisigTxnsResponse)
-    })
-    it("should return create account transaction", async function () {
-      const mockCall = jest.fn(async () => {
-        return mockLedgerListAccountCreateResponse
-      })
-      const ledger = setupLedger(mockCall)
-      const res = await ledger.list()
-      expect(res).toEqual(expectedMockLedgerListAccountCreateResponse)
-    })
-    it("should be called with correct args", async () => {
-      const mockCall = jest.fn(async () => {
-        return mockLedgeListResponseMessage
-      })
-      const ledger = setupLedger(mockCall)
-      await ledger.list()
-      expect(mockCall).toHaveBeenCalledWith(
-        "ledger.list",
-        // @ts-ignore
-        new Map([
-          [0, 10],
-          [1, 2],
-          [2, new Map()],
-        ]),
-      )
-
-      mockCall.mockReset()
-      mockCall.mockResolvedValueOnce(mockLedgeListResponseMessage)
-      const txnId = new Uint8Array(Buffer.from("txn"))
-      await ledger.list({
-        order: OrderType.ascending,
-        count: 20,
-        filters: {
-          symbols: "symbol1",
-          accounts: "account1",
-          txnIdRange: [
-            undefined,
-            { boundType: BoundType.exclusive, value: txnId },
-          ],
-        },
-      })
-      expect(mockCall).toHaveBeenCalledWith(
-        "ledger.list",
-        // @ts-ignore
-        new Map([
-          [0, 20],
-          [1, OrderType.ascending],
-          [
-            2,
-            // @ts-ignore
-            new Map([
-              [0, "account1"],
-              [2, "symbol1"],
-              [3, new Map([[1, [1, txnId]]])],
-            ]),
-          ],
-        ]),
-      )
-    })
-  })
-  describe("makeListFilters", () => {
-    it("should construct the filters", () => {
-      const txnId = new Uint8Array(Buffer.from("txnid1"))
-      const accounts = "account1"
-      const symbols = "symbol1"
-      const filters = makeListFilters({
-        accounts,
-        symbols,
-        txnIdRange: [
-          undefined,
-          { boundType: BoundType.exclusive, value: txnId },
-        ],
-      })
-      expect(filters).toEqual(
-        //@ts-ignore
-        new Map([
-          [0, accounts],
-          [2, symbols],
-          [3, new Map([[1, [1, txnId]]])],
-        ]),
-      )
-
-      const filters2 = makeListFilters({
-        accounts,
-        symbols,
-        txnIdRange: [
-          { boundType: BoundType.inclusive, value: txnId },
-          undefined,
-        ],
-      })
-      expect(filters2).toEqual(
-        // @ts-ignore
-        new Map([
-          [0, accounts],
-          [2, symbols],
-          [3, new Map([[0, [0, txnId]]])],
-        ]),
-      )
-    })
-  })
-  describe("setRangeBound", () => {
-    it("should make bound range", () => {
-      let map = new Map()
-      setRangeBound({
-        rangeMap: map,
-        rangeType: RangeType.lower,
-        boundType: BoundType.exclusive,
-        value: "123",
-      })
-      expect(map).toEqual(new Map([[0, [1, "123"]]]))
-
-      map = new Map()
-      setRangeBound({
-        rangeMap: map,
-        rangeType: RangeType.upper,
-        boundType: BoundType.inclusive,
-        value: "123",
-      })
-      expect(map).toEqual(new Map([[1, [0, "123"]]]))
-
-      map = new Map()
-      setRangeBound({
-        rangeMap: map,
-        rangeType: RangeType.upper,
-        boundType: BoundType.unbounded,
-        value: "123",
-      })
-      expect(map).toEqual(new Map([[1, []]]))
     })
   })
 })
