+++ src/utils/index.ts
@@ -26,7 +26,11 @@ export function applyMixins<N extends Network, M extends NetworkModule[]>(
 
 export function throwOnErrorResponse(msg: Message) {
   const content = msg.getContent()?.get(4)
-  if (content instanceof Map && content.get(0) === -1) {
+  if (
+    content instanceof Map &&
+    typeof content.get(0) === "number" &&
+    Math.sign(content.get(0)) === -1
+  ) {
     throw new ManyError(Object.fromEntries(content) as SerializedManyError)
   }
   return msg
